chore: moved login page and assets
