+++ src/Router.tsx
@@ -4,7 +4,7 @@ import Budgets from './pages/Budgets'
 import Dashboard from './pages/Dashboard'
 
 import Home from './pages/Home'
-import Login from './pages/Login'
+import Login from './pages/Login/Login'
 
 export function Router() {
   return (
