+++ src/components/HomeAvatar.tsx
@@ -1,3 +1,3 @@
 export default function HomeAvatar() {
-  return <img src="/public/ju-logo.png" className="rounded-full w-40 h-40" />
+  return <img src="ju-logo.png" className="rounded-full w-40 h-40" />
 }
